Fix empty CTCP ACTION sending two literal underscores on Discord
Wrapping of IRC CTCP ACTION commands with underscores to mimic the italicization on Discord happens before the logic for ZWB treatment of empty messages. This means that we send two underscore characters for an empty CTCP ACTION, which renders as literal underscores in the Discord UI (instead of the intended empty message). 

This commit fixes it by not wrapping a CTCP ACTION with underscores if it is empty. Having any non-zero whitespace in the CTCP ACTION will not cause this issue (only the true empty string), as Discord renders the sequence `_ _` as empty but `__` as two underscores.